DEV Change behaviour of 'where' command
When called without argument, it now prints the location pointed
to by the key 'here' instead of listing all keys and locations.

The rationale is that a list of all keys can be obtained by
typing 'where' and pressing tab to trigger autocompletion.
Furthermore this way the behaviour is symmetric between 'here'
and 'where'.